travel_planning_agent/pyproject.toml
[project]
name = "evaluation-agent"
version = "0.1.0"
description = "Evaluation_Agent"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5",
    "google-adk>=1.3.0",
    "python-dotenv",
    "uvicorn",
] 
